requirements.txt
transformers>=4.39.0,<5.0.0
peft>=0.10.0,<1.0.0
bitsandbytes>=0.43.0,<1.0.0
datasets>=2.18.0,<3.0.0
torch>=2.1.0,<3.0.0
tqdm
evaluate
scipy
accelerate>=0.28.0,<1.0.0
psutil
pyyaml
scikit-learn
huggingface_hub>=0.20.0
python-dotenv>=1.0.0

# Development dependencies
# pip install pytest  (for running tests)
